+++ packages/components-react/projects/nextjs/pages/overview.tsx
@@ -1,5 +1,5 @@
 /* Auto Generated File */
-import { PAccordion, PBanner, PButton, PButtonGroup, PButtonPure, PCarousel, PCheckboxWrapper, PContentWrapper, PDivider, PFieldsetWrapper, PFlex, PFlexItem, PGrid, PGridItem, PHeadline, PIcon, PInlineNotification, PLink, PLinkPure, PLinkSocial, PMarque, PorscheDesignSystemProvider, PPagination, PPopover, PRadioButtonWrapper, PSegmentedControl, PSegmentedControlItem, PSelectWrapper, PSpinner, PStepperHorizontal, PStepperHorizontalItem, PTable, PTableBody, PTableCell, PTableHead, PTableHeadCell, PTableHeadRow, PTableRow, PTabs, PTabsBar, PTabsItem, PTag, PTagDismissible, PText, PTextareaWrapper, PTextFieldWrapper, PTextList, PTextListItem } from '@porsche-design-system/components-react';
+import { PAccordion, PBanner, PButton, PButtonGroup, PButtonPure, PCarousel, PCheckboxWrapper, PContentWrapper, PDivider, PFieldsetWrapper, PFlex, PFlexItem, PGrid, PGridItem, PHeadline, PIcon, PInlineNotification, PLink, PLinkPure, PLinkSocial, PLinkTile, PMarque, PorscheDesignSystemProvider, PPagination, PPopover, PRadioButtonWrapper, PSegmentedControl, PSegmentedControlItem, PSelectWrapper, PSpinner, PStepperHorizontal, PStepperHorizontalItem, PTable, PTableBody, PTableCell, PTableHead, PTableHeadCell, PTableHeadRow, PTableRow, PTabs, PTabsBar, PTabsItem, PTag, PTagDismissible, PText, PTextareaWrapper, PTextFieldWrapper, PTextList, PTextListItem } from '@porsche-design-system/components-react';
 import { useEffect, useState } from 'react';
 import type { NextPage } from 'next';
 import { componentsReady } from '@porsche-design-system/components-react';
@@ -57,6 +57,10 @@ const OverviewPage: NextPage = (): JSX.Element => {
     p-banner,
     my-prefix-p-banner {
       --p-banner-position-type: static;
+    }
+    p-link-tile,
+    my-prefix-p-link-tile {
+      max-width: 400px;
     }
   `;
 
@@ -282,6 +286,12 @@ const OverviewPage: NextPage = (): JSX.Element => {
               <div>Slide 3</div>
             </PCarousel>
           </div>
+
+          <div className="playground light" title="should render default link tile">
+            <PLinkTile href="#" label="Some Label" description="Default">
+              <img src="/static/porsche_beach.jpg" alt="Beach" />
+            </PLinkTile>
+          </div>
         </div>
 
         <PorscheDesignSystemProvider prefix="my-prefix">
@@ -506,6 +516,12 @@ const OverviewPage: NextPage = (): JSX.Element => {
                 <div>Slide 2</div>
                 <div>Slide 3</div>
               </PCarousel>
+            </div>
+
+            <div className="playground light" title="should render default link tile with custom prefix">
+              <PLinkTile href="#" label="Some Label" description="Default">
+                <img src="/static/porsche_beach.jpg" alt="Beach" />
+              </PLinkTile>
             </div>
           </div>
         </PorscheDesignSystemProvider>