+++ packages/storefront/storefront.config.ts
@@ -203,6 +203,14 @@ export const config: StorefrontConfig = {
       // @ts-ignore
       Props: [() => import('@/../../components/src/components/link-social/link-social.props.md')],
     },
+    'Link Tile': {
+      // @ts-ignore
+      Examples: [() => import('@/../../components/src/components/link-tile/link-tile.examples.md')],
+      // @ts-ignore
+      Usage: [() => import('@/../../components/src/components/link-tile/link-tile.usage.md')],
+      // @ts-ignore
+      Props: [() => import('@/../../components/src/components/link-tile/link-tile.props.md')],
+    },
     Marque: {
       // @ts-ignore
       Examples: [() => import('@/../../components/src/components/marque/marque.examples.md')],